+++ packages/gcf-utils/src/gcf-utils.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
-import {createProbot, Probot, ProbotOctokit, Options} from 'probot';
+import {createProbot, Probot, Options} from 'probot';
 import {ApplicationFunction} from 'probot/lib/types';
 import {createProbotAuth} from 'octokit-auth-probot';
 
@@ -33,8 +33,6 @@ import {v4} from 'uuid';
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const LoggingOctokitPlugin = require('../src/logging/logging-octokit-plugin.js');
 const RUNNING_IN_TEST = process.env.NODE_ENV === 'test';
-
-type ProbotOctokitType = InstanceType<typeof ProbotOctokit>;
 
 interface Scheduled {
   repo?: string;
@@ -91,16 +89,31 @@ export const getCommentMark = (installationId: number): string => {
   return `<!-- probot comment [${installationId}]-->`;
 };
 
+/**
+ * It creates a comment, or if the bot already created a comment, it
+ * updates the same comment.
+ *
+ * @param {Octokit} octokit - The Octokit instance.
+ * @param {string} owner - The owner of the issue.
+ * @param {string} repo - The name of the repository.
+ * @param {number} issueNumber - The number of the issue.
+ * @param {number} installationId - A unique number for identifying the issue
+ *   comment.
+ * @param {string} commentBody - The body of the comment.
+ * @param {boolean} onlyUpdate - If set to true, it will only update an
+ *   existing issue comment.
+ */
 export const addOrUpdateIssueComment = async (
-  github: ProbotOctokitType,
+  octokit: Octokit,
   owner: string,
   repo: string,
   issueNumber: number,
   installationId: number,
-  commentBody: string
+  commentBody: string,
+  onlyUpdate = false
 ) => {
   const commentMark = getCommentMark(installationId);
-  const listCommentsResponse = await github.issues.listComments({
+  const listCommentsResponse = await octokit.issues.listComments({
     owner: owner,
     repo: repo,
     per_page: 50, // I think 50 is enough, but I may be wrong.
@@ -110,7 +123,7 @@ export const addOrUpdateIssueComment = async (
   for (const comment of listCommentsResponse.data) {
     if (comment.body?.includes(commentMark)) {
       // We found the existing comment, so updating it
-      await github.issues.updateComment({
+      await octokit.issues.updateComment({
         owner: owner,
         repo: repo,
         comment_id: comment.id,
@@ -119,8 +132,8 @@ export const addOrUpdateIssueComment = async (
       found = true;
     }
   }
-  if (!found) {
-    await github.issues.createComment({
+  if (!found && !onlyUpdate) {
+    await octokit.issues.createComment({
       owner: owner,
       repo: repo,
       issue_number: issueNumber,
@@ -210,9 +223,12 @@ export class GCFBootstrapper {
    * Parse the event name, delivery id, signature and task id from the request headers
    * @param request incoming trigger request
    */
-  private static parseRequestHeaders(
-    request: express.Request
-  ): {name: string; id: string; signature: string; taskId: string} {
+  private static parseRequestHeaders(request: express.Request): {
+    name: string;
+    id: string;
+    signature: string;
+    taskId: string;
+  } {
     const name =
       request.get('x-github-event') || request.get('X-GitHub-Event') || '';
     const id =
@@ -260,9 +276,8 @@ export class GCFBootstrapper {
       this.probot =
         this.probot || (await this.loadProbot(appFn, wrapOptions?.logging));
 
-      const {name, id, signature, taskId} = GCFBootstrapper.parseRequestHeaders(
-        request
-      );
+      const {name, id, signature, taskId} =
+        GCFBootstrapper.parseRequestHeaders(request);
 
       const triggerType: TriggerType = GCFBootstrapper.parseTriggerType(
         name,
@@ -380,19 +395,26 @@ export class GCFBootstrapper {
           },
         }
       );
+      const promises: Array<Promise<void>> = new Array<Promise<void>>();
+      const batchNum = 30;
       for await (const response of installationsPaginated) {
         for (const repo of response.data) {
           if (repo.archived === true || repo.disabled === true) {
             continue;
           }
-          await this.scheduledToTask(
-            repo.full_name,
-            id,
-            body,
-            eventName,
-            signature
+          promises.push(
+            this.scheduledToTask(repo.full_name, id, body, eventName, signature)
           );
+          if (promises.length >= batchNum) {
+            await Promise.all(promises);
+            promises.splice(0, promises.length);
+          }
         }
+      }
+      // Wait for the rest.
+      if (promises.length > 0) {
+        await Promise.all(promises);
+        promises.splice(0, promises.length);
       }
     }
   }
@@ -456,9 +478,11 @@ export class GCFBootstrapper {
   }
 
   private parseRequestBody(req: express.Request): Scheduled {
-    let body = (Buffer.isBuffer(req.body)
-      ? JSON.parse(req.body.toString('utf8'))
-      : req.body) as Scheduled;
+    let body = (
+      Buffer.isBuffer(req.body)
+        ? JSON.parse(req.body.toString('utf8'))
+        : req.body
+    ) as Scheduled;
     // PubSub messages have their payload encoded in body.message.data
     // as a base64 blob.
     if (body.message && body.message.data) {