+++ packages/gcf-utils/test/gcf-bootstrapper.ts
@@ -53,10 +53,8 @@ describe('GCFBootstrapper', () => {
       response,
       'send'
     );
-    const sendStatusStub: sinon.SinonStub<
-      [number],
-      express.Response
-    > = sinon.stub(response, 'sendStatus');
+    const sendStatusStub: sinon.SinonStub<[number], express.Response> =
+      sinon.stub(response, 'sendStatus');
 
     let req: express.Request;
 
@@ -71,9 +69,12 @@ describe('GCFBootstrapper', () => {
       req = express.request;
 
       bootstrapper = new GCFBootstrapper();
-      configStub = sinon
-        .stub(bootstrapper, 'getProbotConfig')
-        .resolves({appId: 1234, secret: 'foo', webhookPath: 'bar'});
+      configStub = sinon.stub(bootstrapper, 'getProbotConfig').resolves({
+        appId: 1234,
+        secret: 'foo',
+        webhookPath: 'bar',
+        privateKey: 'cert',
+      });
       // This replaces the authClient with an auth client that uses an
       // API Key, this ensures that we will not attempt to lookup application
       // default credentials:
@@ -404,9 +405,12 @@ describe('GCFBootstrapper', () => {
 
     beforeEach(() => {
       bootstrapper = new GCFBootstrapper();
-      configStub = sinon
-        .stub(bootstrapper, 'getProbotConfig')
-        .resolves({appId: 1234, secret: 'foo', webhookPath: 'bar'});
+      configStub = sinon.stub(bootstrapper, 'getProbotConfig').resolves({
+        appId: 1234,
+        secret: 'foo',
+        webhookPath: 'bar',
+        privateKey: 'cert',
+      });
     });
 
     afterEach(() => {